ensure bfloat latents (fix uint8 cast bug)

+++ owl_wms/trainers/world_trainer.py
@@ -202,7 +202,7 @@ class WorldTrainer(BaseTrainer):
         if "rgb" in batch:
             assert "x" not in batch, "passed rgb to convert, but already have batch item `x` (latents)"
             raw_rgb = batch.pop("rgb")
-            batch["x"] = self.encoder_decoder.encode(raw_rgb).type_as(raw_rgb)
+            batch["x"] = self.encoder_decoder.encode(raw_rgb).bfloat16()
         return batch
 
     @torch.no_grad()
